Removed sidebar for normal users and created footer
